bq27xxx_battery: do not update cached flags prematurely
Commit 243f8ffc883a1 ("power: supply: bq27xxx_battery: Notify also about status
changes") intended to notify userspace when the status changes, based on the
flags register. However, the cached state is updated too early, so the test for
the flags is always false. Remove the premature update, to update the flags
only after the condition is evaluated, when the entire cache is refreshed.

Fixes: 243f8ffc883a1 ("power: supply: bq27xxx_battery: Notify also about status changes")
Signed-off-by: Sicelo A. Mhlongo <absicsz@gmail.com>